// dfs/ui/src/types/fs/filesystem.ts
export interface FileSystemObject {
    /**
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * @generated from protobuf oneof: type
     */
    type: {
        oneofKind: "file";
        /**
         * @generated from protobuf field: fs.File file = 3;
         */
        file: File;
    } | {
        oneofKind: "directory";
        /**
         * @generated from protobuf field: fs.Directory directory = 4;
         */
        directory: Directory;
    } | {
        oneofKind: "link";
        /**
         * @generated from protobuf field: fs.Link link = 5;
         */
        link: Link;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message fs.File
 */
export interface File {
    /**
     * @generated from protobuf field: int64 size = 1;
     */
    size: bigint;
    /**
     * @generated from protobuf field: fs.ChunkHashes hashes = 2;
     */
    hashes?: ChunkHashes;
}
/**
 * @generated from protobuf message fs.ChunkHashes
 */
export interface ChunkHashes {
    /**
     * @generated from protobuf field: int32 chunkSize = 1;
     */
    chunkSize: number;
    /**
     * @generated from protobuf field: repeated bytes hash = 2;
     */
    hash: Uint8Array[];
}
/**
 * @generated from protobuf message fs.Directory
 */
export interface Directory {
    /**
     * @generated from protobuf field: repeated bytes entries = 1;
     */
    entries: Uint8Array[];
}
/**
 * @generated from protobuf message fs.Link
 */
export interface Link {
    /**
     * @generated from protobuf field: string target_path = 1;
     */
    targetPath: string; // (for now) we can't target hashes in some cases (e.g., linking to ancestor causes a cycle), so this will have to do
}
/**
 * @generated from protobuf message fs.ObjectWithHash
 */
export interface ObjectWithHash {
    /**
     * @generated from protobuf field: fs.FileSystemObject object = 1;
     */
    object?: FileSystemObject;
    /**
     * @generated from protobuf field: bytes hash = 2;
     */
    hash: Uint8Array;
}
/**
 * @generated from protobuf message fs.ObjectList
 */
export interface ObjectList {
    /**
     * @generated from protobuf field: repeated fs.ObjectWithHash data = 1;
     */
    data: ObjectWithHash[];
}
// @generated message type with reflection information, may provide speed optimized methods
class FileSystemObject$Type extends MessageType<FileSystemObject> {
    constructor() {
        super("fs.FileSystemObject", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "file", kind: "message", oneof: "type", T: () => File },
            { no: 4, name: "directory", kind: "message", oneof: "type", T: () => Directory },
            { no: 5, name: "link", kind: "message", oneof: "type", T: () => Link }
        ]);
    }
    create(value?: PartialMessage<FileSystemObject>): FileSystemObject {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.type = { oneofKind: undefined };
        if (value !== undefined)
            reflectionMergePartial<FileSystemObject>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: FileSystemObject): FileSystemObject {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* fs.File file */ 3:
                    message.type = {
                        oneofKind: "file",
                        file: File.internalBinaryRead(reader, reader.uint32(), options, (message.type as any).file)
                    };
                    break;
                case /* fs.Directory directory */ 4:
                    message.type = {
                        oneofKind: "directory",
                        directory: Directory.internalBinaryRead(reader, reader.uint32(), options, (message.type as any).directory)
                    };
                    break;
                case /* fs.Link link */ 5:
                    message.type = {
                        oneofKind: "link",
                        link: Link.internalBinaryRead(reader, reader.uint32(), options, (message.type as any).link)
                    };
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: FileSystemObject, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* fs.File file = 3; */
        if (message.type.oneofKind === "file")
            File.internalBinaryWrite(message.type.file, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* fs.Directory directory = 4; */
        if (message.type.oneofKind === "directory")
            Directory.internalBinaryWrite(message.type.directory, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        /* fs.Link link = 5; */
        if (message.type.oneofKind === "link")
            Link.internalBinaryWrite(message.type.link, writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message fs.FileSystemObject
 */
export const FileSystemObject = new FileSystemObject$Type();
// ...
export const File = new File$Type();
// ...
export const Directory = new Directory$Type();
// ...
export const Link = new Link$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ObjectWithHash$Type extends MessageType<ObjectWithHash> {
    constructor() {
        super("fs.ObjectWithHash", [
            { no: 1, name: "object", kind: "message", T: () => FileSystemObject },
            { no: 2, name: "hash", kind: "scalar", T: 12 /*ScalarType.BYTES*/ }
        ]);
    }
    create(value?: PartialMessage<ObjectWithHash>): ObjectWithHash {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.hash = new Uint8Array(0);
        if (value !== undefined)
            reflectionMergePartial<ObjectWithHash>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ObjectWithHash): ObjectWithHash {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* fs.FileSystemObject object */ 1:
                    message.object = FileSystemObject.internalBinaryRead(reader, reader.uint32(), options, message.object);
                    break;
                case /* bytes hash */ 2:
                    message.hash = reader.bytes();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ObjectWithHash, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* fs.FileSystemObject object = 1; */
        if (message.object)
            FileSystemObject.internalBinaryWrite(message.object, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* bytes hash = 2; */
        if (message.hash.length)
            writer.tag(2, WireType.LengthDelimited).bytes(message.hash);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message fs.ObjectWithHash
 */
export const ObjectWithHash = new ObjectWithHash$Type();